NKSR Sparse Feature Hierarchy (#8)
This PR introduces the lowest level of foundation classes for NKSR.

The coordinate transform API associated with fVDB's GridBatch is less
explicit than we'd like it to be, so the first level of functionality
added to this PR is the CoordXform class, which represents an
abstraction that can transform tensors of points, jagged tensors of
points, and also the same for bounding boxes and jagged tensors of
bounding boxes. The xforms are composable, and a basic scale/translate
xform that matches the scale & offset needed by NKSR is added. See the
file `surface_reconstruction/tests/unit/test_coord_xform.py` for
extensive testing and usage examples.

The SparseFeatureHierarchy is defined in the file
`surface_reconstruction/nksr/nksr_fvdb/sparse_feature_hierarchy.py`, and
is composed of a immutable list of SparseFeatureLevels ordered from fine
to coarse. A SparseFeatureLevel is just a pairing of a coordinate
transform with a grid batch. Care is taken to make sure that each grid
in the batch can potentially have separate transforms, though the code
does assume (or want, at least) a single, consistent voxel size for each
batch item in the level hierarchy. Factory functions demonstrate how the
hierarchy might be built.

We need decent synthetic test data for testing these foundational
components, so there's a fairly robust set of fake scene generation
tools inside the file `surface_reconstruction/tests/unit/utils.py`.
These are separately unit tested.

---------

Signed-off-by: Christopher Horvath <chorvath@nvidia.com>